+++ toolium/driver_wrappers_pool.py
@@ -121,10 +121,6 @@ class DriverWrappersPool(object):
             if not driver_wrapper.driver:
                 continue
             try:
-                # Delete IE cookies (Workaround of IEDriverServer bug)
-                driver_type = driver_wrapper.config.get('Driver', 'type')
-                if not maintain_default and driver_type.split('-')[0] in ['iexplore', 'edge']:
-                    driver_wrapper.driver.delete_all_cookies()
                 # Stop driver
                 driver_wrapper.driver.quit()
                 # Download video if necessary