+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.2.2
 *In development*
 
 - Fix error comparing screenshots in mobile tests
+- Fix image size when enlarging a vertical image in visual testing reports
 
 v1.2.1
 ------